refactor: `Calculator` component
