+++ src/renderer/common/utils/dynamic-calculate.ts
