+++ src/renderer/common/utils/get-last-button.ts
@@ -1,0 +1,10 @@
+import { IButton } from '../../types/interfaces';
+
+export const getLastElem = (query: IButton[]): IButton => {
+  if (query.length === 0)
+    return {
+      char: '',
+      operation: false,
+    };
+  else return query[query.length - 1];
+};
