+++ src/renderer/common/utils/get-query-string.ts
@@ -1,0 +1,5 @@
+import { IButton } from '../../types/interfaces';
+
+export const getQueryString = (query: IButton[]): string => {
+  return query.map((elem) => elem.char).join('');
+};
