+++ src/renderer/common/utils/operation-exists.ts
@@ -1,0 +1,8 @@
+import { IButton } from '../../types/interfaces';
+
+export const operationExists = (query: IButton[]): boolean => {
+  for (const item of query) {
+    if (item.operation) return true;
+  }
+  return false;
+};
