+++ src/renderer/common/utils/updated-query.ts
@@ -1,0 +1,7 @@
+export const updatedQuery = (result: string) => {
+  if (result === 'Error!') return [];
+  const res = result.split('').map((char) => {
+    return { char, operation: false };
+  });
+  return res;
+};
