+++ src/renderer/types/enums.ts
