+++ src/renderer/types/interfaces.ts
